Clean code, some type errors
