+++ src/main/scala/calc/Compiler.scala
@@ -4,6 +4,7 @@ import org.scalajs.core.ir
 import ir.{Trees => irt, Types => irtpe}
 import ir.Definitions._
 import scala.collection.immutable.Map
+import java.util.UUID.randomUUID
 
 /** Main compiler.
  *
@@ -14,9 +15,9 @@ object Compiler {
 
   private final val MainClassFullName = MainObjectFullName + "$"
 
-  implicit val defaultEnv = Map[String, irt.Tree] ()
-  implicit val defaultTypeEnv = Map[String, irtpe.Type]()
-  implicit val defaultParams = Set[String]()
+  //implicit val defaultEnv = Map[String, irt.Tree] ()
+  //implicit val defaultTypeEnv = Map[String, irtpe.Type]()
+  //implicit val defaultParams = Set[String]()
   /** Compile an expression tree into a full `ClassDef`.
    *
    *  You do not need to modify this method.
@@ -37,7 +38,7 @@ object Compiler {
                 irtpe.NoType),
             irt.StoreModule(classType, irt.This()(classType)))))(
         irt.OptimizerHints.empty, None)
-    val body = compileExpr(tree)
+    val body = compileExpr(tree)()
     val methodDef = irt.MethodDef(static = false,
         irt.Ident("main__D", Some("main")), Nil, irtpe.DoubleType, body)(
         irt.OptimizerHints.empty, None)
@@ -73,25 +74,23 @@ object Compiler {
       case Let(i, v, body) => findCapture(body) ++ findCapture(v) - i.name
       case Ident(n) => Set(n)
       case If(cond, thenp, elsep) => findCapture(cond) ++ findCapture(thenp) ++ findCapture(elsep)
-      case Closure(idents, body) => findCapture(body) -- idents.toSet.flatMap((x:Tree) => findCapture(x))
-      case Call(f, args) => findCapture(f) ++ args.toSet.flatMap((x:Tree) => findCapture(x))
+      case Closure(idents, body) => findCapture(body) -- idents.flatMap((x:Tree) => findCapture(x)).toSet
+      case Call(f, args) => findCapture(f) ++ args.flatMap((x:Tree) => findCapture(x)).toSet
     }
   }
 
-  def compileClosure(tree: Closure, ident: String)(implicit env:Map[String, irt.Tree],typeEnv:Map[String, irtpe.Type]): irt.Tree ={
-    val captureSet = findCapture(tree) - ident
-    val captureParam = captureSet.map(s => irt.ParamDef(irt.Ident(s)(tree.pos), typeEnv(s), false, false)(tree.pos)).toList
-    val params = tree.params.map(s => irt.ParamDef(irt.Ident(s.name)(tree.pos), irtpe.AnyType, false, false)(tree.pos))
-    irt.Closure(captureParam, params, compileExpr(tree.body)(env, typeEnv), captureSet.map(s => env(s)).toList)(tree.pos)
+  var seed = 1
+  def generateName():String = {
+    seed += 1
+    "generated_"+seed.toString
   }
-
 
   /** Compile an expression tree into an IR `Tree`, which is an expression
    *  that evaluates to the result of the tree.
    *
    *  This is the main method you have to implement.
    */
-  def compileExpr(tree: Tree)(implicit env:Map[String, irt.Tree],typeEnv:Map[String, irtpe.Type]): irt.Tree = {
+  def compileExpr(tree: Tree)(env:Map[String, irt.Tree] = Map(),typeEnv:Map[String, irtpe.Type] = Map()): irt.Tree = {
     // TODO
     implicit val pos = tree.pos
 
@@ -106,15 +105,10 @@ object Compiler {
           case "*" => irt.BinaryOp.Double_*
           case "/" => irt.BinaryOp.Double_/
         }
-        irt.BinaryOp(jsop, irt.Unbox(compileExpr(lhs)(env, typeEnv),'D'), irt.Unbox(compileExpr(rhs)(env, typeEnv),'D'))
+        irt.BinaryOp(jsop, compileExpr(lhs)(env, typeEnv), compileExpr(rhs)(env, typeEnv))
       }
 
       case Let(ident, value, body) => {
-        /*var jsV:irt.Tree = if (value.isInstanceOf[Closure]){
-          compileClosure(value.asInstanceOf[Closure], ident.name)(env, typeEnv)
-        } else{
-          compileExpr(value)(env, typeEnv)
-        }*/
         val jsV = compileExpr(value)(env, typeEnv)
         val jsIdent = irt.Ident(ident.name)
         val jsId = irt.VarDef(jsIdent, jsV.tpe, false, jsV)
@@ -123,10 +117,12 @@ object Compiler {
 
       case Ident(n) => {
         try {
-          irt.VarRef(irt.Ident(n))(typeEnv(n))
+          typeEnv(n) match {
+            case irtpe.DoubleType => irt.Unbox(irt.VarRef(irt.Ident(n))(typeEnv(n)),'D')
+            case _ => irt.VarRef(irt.Ident(n))(typeEnv(n))
+          }
         }catch{
-          case e:Exception => irt.VarRef(irt.Ident(n))(irtpe.AnyType)
-          //throw new Exception(s"The variable ${n} is not in the scope")
+          case e:Exception => irt.VarRef(irt.Ident(n)) (irtpe.DoubleType)
         }
       }
 
@@ -142,25 +138,16 @@ object Compiler {
       }
 
       case Closure(idList, body) => {
-        /*val capturePara = idList.map(s => s match{
-          case Ident(n) => irt.ParamDef(irt.Ident(n), irtpe.AnyType, false, false)
-          case _ => throw new Exception("Function def error!")
-        })*/
-        val captureSet = findCapture(tree)
-        val captureParam = captureSet.map(s => irt.ParamDef(irt.Ident(s), typeEnv(s), false, false)).toList
+        val captureList = findCapture(tree).toList
+        val captureParam = captureList.map(s => irt.ParamDef(irt.Ident(s), typeEnv(s), false, false))
         val params = idList.map(s => irt.ParamDef(irt.Ident(s.name), irtpe.AnyType, false, false))
-        irt.Closure(captureParam, params, compileExpr(body)(env, typeEnv), captureSet.map(s => env(s)).toList)
+        val newTypeEnv = idList.foldLeft(typeEnv) ((env:Map[String,irtpe.Type], s:Ident) => env+(s.name -> irtpe.DoubleType))
+        irt.Closure(captureParam, params, compileExpr(body)(env,newTypeEnv), captureList.map(s => env(s)))
       }
 
       case Call(f, args) => {
         val jsFunction = compileExpr(f)(env, typeEnv)
-        /*val jsClass = irt.GetClass(jsFunction)
-        val fName = f match {
-          case Ident(n) => n
-          case _ => throw new Exception("TO")
-        }*/
-        //irt.Apply(irt.This()(irtpe.ClassType(MainClassFullName)), irt.Ident(fName), args.map(s => compileExpr(s)))(irtpe.DoubleType)
-        irt.Unbox(irt.JSFunctionApply(jsFunction,args.map(s => compileExpr(s)(env, typeEnv))),'D')
+        irt.Unbox(irt.JSFunctionApply(jsFunction,args.map(s => irt.Unbox(compileExpr(s)(env, typeEnv),'D'))),'D')
       }
